Export bar scrolls to top when switching to 'All Lectures'
Find and Replace view can be called and displayed but lacks functionality in this commit.

The printableSchedule() func in SidebarVC now accurately produces a textual representation of a courses week schedule.

Lecture TextView "out of range" bug has been fixed and the function that follows a selection change has been tweaked.
	• Removed couple unnecessary lines and now understand effectiveRange parameter.
	-> It's a pointer so its being changed by the function for you to be able to see how far along the attribute applies.

SidebarVC will select course automatically if there is only 1 course in a semester.